+++ lib/UserFilter.php
@@ -20,18 +20,18 @@
 
 namespace Opis\HttpRouting;
 
-use Opis\Routing\Contracts\RouteInterface;
-use Opis\Routing\Contracts\FilterInterface;
-use Opis\Routing\Contracts\PathInterface;
+use Opis\Routing\Path;
+use Opis\Routing\Route;
+use Opis\Routing\FilterInterface;
 
 class UserFilter implements FilterInterface
 {
    
-    public function pass(PathInterface $path, RouteInterface $route)
+    public function pass(Path $path, Route $route)
     {
         $filters = $route->getFilters();
         
-        foreach($route->get('prefilter', array()) as $name)
+        foreach($route->get('beforefilter', array()) as $name)
         {
             if(isset($filters[$name]))
             {